// src/types/database.ts
export type Post = {
  id: string;
  title: string;
  description: string;
  category: Category;
  author: Author;
  slug: string;
  image: string;
  body: string;
  date_created: string;
  date_updated: string;
};

export type Category = {
  id: string;
  title: string;
  slug?: string;
  description?: string;
};

export type Author = {
  id: string;
  first_name: string;
  last_name: string;
};